Refactor strict compiler warnings
Fixes compiler warnings.

Sets strict compiler flags only for non-generated non-external sources.

Removes "-Werror" compiler flag so that compilation doesn't stop after a
warning. The flag makes development very difficult. Occasionally, it's a
good idea to check CI build output -- GitHub Actions list the warnings
very well.

+++ src/gui/icon_list.h
@@ -1,7 +1,7 @@
 // This file is generated with "update_icon_font.py" from FontAwesome's metadata.
 
 struct Icon {
-    unsigned int unicode;
+    unsigned short unicode;
     bool isBrand;
     const char *searchTerms;
 };
